requirements.txt
numpy>=1.24.0
pywebview>=4.4.1